remove Map class
The Map class wasn't doing anything the Room class wasn't doing for
itself.  Refactor Game.__init__ to accept a list of rooms directly.

Game doesn't really need the list of rooms either.  It just needs to
know the starting location, which in the future, can be either passed
into Game.__init__ or require that the Player be instantiated with a
starting location.